Adding some modification at the parsing part

+++ includes/philo.h
@@ -6,7 +6,7 @@
 /*   By: souel-bo <souel-bo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/22 10:11:55 by souel-bo          #+#    #+#             */
-/*   Updated: 2025/02/22 11:57:35 by souel-bo         ###   ########.fr       */
+/*   Updated: 2025/02/22 16:11:20 by souel-bo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,15 +16,16 @@
 #include <stdio.h>
 #include <unistd.h>
 #include <pthread.h>
+#include <stdlib.h>
 
 typedef struct s_philo
 {
-    //number_of_philosophers time_to_die time_to_eat time_to_sleep [number_of_times_each_philosopher_must_eat]
-    int philos;
-    int death;
-    int eat;
-    int sleep;
-    int meals;
+    //number_of_philosophers//////time_to_die/////time_to_eat/////////time_to_sleep/////[number_of_times_each_philosopher_must_eat]
+    unsigned long philos;
+    unsigned long death;
+    unsigned long eat;
+    unsigned long sleep;
+    unsigned long meals;
 } t_philo;
 
 unsigned long	ft_atoi(const char *str);
